[FIX] base: duration widget for unsupported babel format
In case of exception of type KeyError during the formatting with Babel
instead to raise a Traceback, we first retry with the en_US locale.

Babel fixes each week new bug of formatting like the one in the test,
but we cannot bump our default babel version since it is not in the
stable Ubuntu 22 so it is a best effort fix that will not hide all bugs
but is better that nothing.

https://github.com/python-babel/babel/pull/827

closes odoo/odoo#155978

X-original-commit: 7cc24364058a6924375802e11fe976e12f0cac6a
Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>
Signed-off-by: Jérémy Kersten <jke@odoo.com>